Revert "Unit::Prefix is int not double"
This reverts commit b3bab9b8718afb12b04f08a9570a064ccfdfb6cb.

+++ src/api/libcellml/units.h
@@ -151,7 +151,9 @@ public:
      * @param multiplier The multiplier.
      * @param id The optional @c std::string id to set for this unit.
      */
-    void addUnit(const std::string &reference, int prefix, double exponent,
+    //void addUnit(const std::string &reference, int prefix, double exponent,
+    //             double multiplier=1.0, const std::string &id="");
+    void addUnit(const std::string &reference, double prefix, double exponent,
                  double multiplier=1.0, const std::string &id="");
 
     /**
@@ -228,7 +230,9 @@ public:
      * @param multiplier The multiplier.
      * @param id The optional @c std::string id to set for this unit.
      */
-    void addUnit(StandardUnit standardRef, int prefix, double exponent,
+    //void addUnit(StandardUnit standardRef, int prefix, double exponent,
+    //             double multiplier=1.0, const std::string &id="");
+    void addUnit(StandardUnit standardRef, double prefix, double exponent,
                  double multiplier=1.0, const std::string &id="");
 
     /**
